+++ packages/kit/src/runtime/server/utils.js
@@ -1,15 +1,7 @@
 /** @import { ServerHooks } from 'types' */
 import * as devalue from 'devalue';
-import { DEV } from 'esm-env';
-import { json, text } from '@sveltejs/kit';
-import { HttpError } from '@sveltejs/kit/internal';
-import { with_request_store } from '@sveltejs/kit/internal/server';
-import { coalesce_to_error, get_message, get_status } from '../../utils/error.js';
-import { negotiate } from '../../utils/http.js';
-import { fix_stack_trace } from '../shared-server.js';
+import { text } from '@sveltejs/kit';
 import { ENDPOINT_METHODS } from '../../constants.js';
-import { escape_html } from '../../utils/escape.js';
-import * as path from '../../utils/path.js';
 
 /**
  * @param {Partial<Record<import('types').HttpMethod, any>>} mod
@@ -47,103 +39,6 @@ export function get_global_name(options) {
 }
 
 /**
- * Return as a response that renders the error.html
- *
- * @param {import('types').SSROptions} options
- * @param {number} status
- * @param {string} message
- */
-export function static_error_page(options, status, message) {
-	let page = options.templates.error({ status, message: escape_html(message) });
-
-	if (__SVELTEKIT_DEV__) {
-		// inject Vite HMR client, for easier debugging
-		page = page.replace('</head>', '<script type="module" src="/@vite/client"></script></head>');
-	}
-
-	return text(page, {
-		headers: { 'content-type': 'text/html; charset=utf-8' },
-		status
-	});
-}
-
-/**
- * @param {import('@sveltejs/kit').RequestEvent} event
- * @param {import('types').RequestState} state
- * @param {import('types').SSROptions} options
- * @param {unknown} error
- */
-export async function handle_fatal_error(event, state, options, error) {
-	error = error instanceof HttpError ? error : coalesce_to_error(error);
-	const body = await handle_error_and_jsonify(event, state, options, error);
-	const status = body.status;
-
-	// ideally we'd use sec-fetch-dest instead, but Safari — quelle surprise — doesn't support it
-	const type = negotiate(event.request.headers.get('accept') || 'text/html', [
-		'application/json',
-		'text/html'
-	]);
-
-	if (event.isDataRequest || type === 'application/json') {
-		return json(body, {
-			status
-		});
-	}
-
-	return static_error_page(options, status, body.message);
-}
-
-/**
- * @param {import('@sveltejs/kit').RequestEvent} event
- * @param {import('types').RequestState} state
- * @param {import('types').SSROptions} options
- * @param {any} error
- * @returns {App.Error | Promise<App.Error>}
- */
-export function handle_error_and_jsonify(event, state, options, error) {
-	if (error instanceof HttpError) {
-		// @ts-expect-error custom user errors may not have a message field if App.Error is overwritten
-		return { message: 'Unknown Error', ...error.body };
-	}
-
-	if (__SVELTEKIT_DEV__ && typeof error == 'object') {
-		fix_stack_trace(error);
-	}
-
-	const status = get_status(error);
-	const message = get_message(error);
-
-	// TODO 4.0 await this, rather than handling the non-Promise case
-	const result = with_request_store({ event, state }, () =>
-		options.hooks.handleError({ error, event, status, message })
-	) ?? { status, message };
-
-	if (result instanceof Promise) {
-		if (!__SVELTEKIT_SUPPORTS_ASYNC__ && state.is_in_render) {
-			console.warn(
-				`To use an async \`handleError\` hook to handle errors that occur during rendering, you must enable \`compilerOptions.experimental.async\` in the SvelteKit plugin of your Vite config. The returned error has been replaced with a generic object`
-			);
-
-			// we're discarding the result, but we still need to prevent an unhandled
-			// rejection if the user's async `handleError` hook rejects
-			result.catch(() => {});
-
-			return {
-				status,
-				message: 'Internal Error'
-			};
-		}
-
-		return result.then((body) => {
-			body ??= { status, message };
-			return { ...body, status: get_status(body, error) };
-		});
-	}
-
-	return { ...result, status: get_status(result, error) };
-}
-
-/**
  * @param {number} status
  * @param {string} location
  */
@@ -163,7 +58,7 @@ export function clarify_devalue_error(event, error) {
 	if (error.path) {
 		return (
 			`Data returned from \`load\` while rendering ${event.route.id} is not serializable: ${error.message} (${error.path}). ` +
-			`If you need to serialize/deserialize custom types, use transport hooks: https://svelte.dev/docs/kit/hooks#Universal-hooks-transport.`
+			`If you need to serialize/deserialize custom types, use transport hooks: https://svelte.dev/docs/kit/hooks#transport.`
 		);
 	}
 
@@ -210,56 +105,6 @@ export function has_prerendered_path(manifest, pathname) {
 		manifest._.prerendered_routes.has(pathname) ||
 		(pathname.at(-1) === '/' && manifest._.prerendered_routes.has(pathname.slice(0, -1)))
 	);
-}
-
-/**
- * Formats the error into a nice message with sanitized stack trace
- * @param {number} status
- * @param {Error} error
- * @param {import('@sveltejs/kit').RequestEvent} event
- */
-export function format_server_error(status, error, event) {
-	const formatted_text = `\n\x1b[1;31m[${status}] ${event.request.method} ${event.url.pathname}\x1b[0m`;
-
-	if (status === 404) {
-		return formatted_text;
-	}
-
-	return `${formatted_text}\n${DEV ? clean_up_stack_trace(error) : error.stack}`;
-}
-
-/**
- * In dev, tidy up stack traces by making paths relative to the current project directory
- * @param {string} file
- */
-let relative = (file) => file;
-
-if (DEV) {
-	try {
-		relative = (file) => path.relative(__SVELTEKIT_ROOT__, file);
-	} catch {
-		// do nothing
-	}
-}
-
-/**
- * Provides a refined stack trace by excluding lines following the last occurrence of a line containing +page. +layout. or +server.
- * @param {Error} error
- */
-export function clean_up_stack_trace(error) {
-	const stack_trace = (error.stack?.split('\n') ?? []).map((line) => {
-		return line.replace(/\((.+)(:\d+:\d+)\)$/, (_, file, loc) => `(${relative(file)}${loc})`);
-	});
-
-	// progressive enhancement for people who haven't configured files.src to something else
-	const last_line_from_src_code = stack_trace.findLastIndex((line) => /\(src[\\/]/.test(line));
-
-	if (last_line_from_src_code === -1) {
-		// default to the whole stack trace
-		return error.stack;
-	}
-
-	return stack_trace.slice(0, last_line_from_src_code + 1).join('\n');
 }
 
 /**